restructured the code. upgraded to python 3.8 created package and moved imports improved documentation
Former-commit-id: 127c2bfaadfb854be2ad7c29dc990f6b996f5e5b

+++ integration_test/Heaviside_column/analytical_wave_prop.py
@@ -40,7 +40,7 @@ def calc_ana_sol(L, K, rho, p0, nb_ele, cycles=10, terms=100):
     # solid discretisation
     H_discre = np.linspace(0, L, nb_ele + 1)
     # time discretisation
-    time = np.linspace(0, (cycles * L / c), np.ceil(c / L)*10)
+    time = np.linspace(0, (cycles * L / c), int(np.ceil(c / L) * 10))
 
     # variable initialisation: u = displacement; p = pressure
     u = np.zeros((H_discre.shape[0], time.shape[0]))
